+++ supreme/src/commonTest/kotlin/at/asitplus/signum/supreme/symmetric/00SymmetricTest.kt
@@ -20,9 +20,9 @@ import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNot
 import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.types.shouldBeInstanceOf
-import kotlinx.datetime.Clock
 import kotlin.random.Random
 import kotlin.random.nextUInt
+import kotlin.time.Clock
 
 @OptIn(HazardousMaterials::class)
 @ExperimentalStdlibApi