+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Dependency Updates:
     * `kotlincrypto:secure-random:0.3.2` -> `kotlincrypto.random:crypto-rand:0.5.0`
         * This fixes key generation in WASM/JS
+* PBKDF2 Support
+* HKDF Support
+* SCrypt Support
 * Add `SpecializedSymmetricEncryptionAlgorithm`
     * This allows `randomKey()` etc to operate on COSE/JWE algorithms
 * Clean up some function signatures: